+++ CHANGELOG.md
@@ -4,7 +4,14 @@ CHANGELOG
 ## Version 1.1 (work in progress)
 
 
+03.05.2013
+- **IMPROVED:** Event base before/after fork hooks support (amal)
+
+27.04.2013
+- **CHANGE:** `Base::$isMaster` refactored to `Base::$hasParent` with inverse value (amal)
+
 25.04.2013
+- **MINOR:** `Base::getTimeForLog` improved (amal)
 - **FEATURE:** `Base::setProcessTitle` now support PHP 5.5 `cli_set_process_title` function (amal)
 - **MINOR:** POSIX (Unix) signals reference improved (amal)